Rename the ConvertToMongoDB attribute
