+++ Specifications/Projections/Builder/Copies/MongoDB/for_ConversionsFromConvertToMongoDBAttributesBuilder/given/all_dependencies.cs
@@ -6,17 +6,17 @@ using Dolittle.SDK.Projections.Copies.MongoDB;
 using Machine.Specifications;
 using Moq;
 
-namespace Dolittle.SDK.Projections.Builder.Copies.MongoDB.for_ConversionsFromMongoDBConvertToAttributesBuilder.given;
+namespace Dolittle.SDK.Projections.Builder.Copies.MongoDB.for_ConversionsFromConvertToMongoDBAttributesBuilder.given;
 
 public class all_dependencies
 {
-    protected static ConversionsFromMongoDBConvertToAttributesBuilder builder;
+    protected static ConversionsFromConvertToMongoDBAttributesBuilder builder;
     protected static ClientBuildResults build_results;
     protected static Mock<IPropertyConversions> conversions;
 
     Establish context = () =>
     {
-        builder = new ConversionsFromMongoDBConvertToAttributesBuilder();
+        builder = new ConversionsFromConvertToMongoDBAttributesBuilder();
         conversions = new Mock<IPropertyConversions>();
         build_results = new ClientBuildResults();
     };
